# Conflicts:
#	Singer.API/ClientApp/package-lock.json
#	Singer.API/ClientApp/package.json
#	Singer.API/ClientApp/src/app/material.module.ts
#	Singer.API/ClientApp/src/app/modules/admin/components/admin-users/admin-list/admin-list.component.ts
#	Singer.API/ClientApp/src/app/modules/admin/components/careusers/overview/overview.component.ts
#	Singer.API/ClientApp/src/app/modules/admin/components/legalguardians/legalguardian-overview/legalguardian-overview.component.ts
#	Singer.API/ClientApp/src/app/modules/dashboard/components/home/home.component.ts
#	Singer.API/ClientApp/src/app/modules/shared/shared.module.ts

+++ Singer.API/Controllers/CareUserController.cs
@@ -4,8 +4,8 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Singer.DTOs.Users;
-using Singer.Models.Users;
 using Singer.Helpers.Exceptions;
+using Singer.Models.Users;
 using Singer.Services.Interfaces;
 
 namespace Singer.Controllers
@@ -15,9 +15,24 @@ namespace Singer.Controllers
    public class CareUserController : DataControllerBase<CareUser, CareUserDTO, CreateCareUserDTO, UpdateCareUserDTO>
    {
       private readonly ICareUserService _careUserService;
-      public CareUserController(ICareUserService careUserService) : base(careUserService)
+      private readonly IDateValidator _dateValidator;
+
+      public CareUserController(ICareUserService careUserService, IDateValidator dateValidator) : base(careUserService)
       {
          _careUserService = careUserService;
+         _dateValidator = dateValidator;
+      }
+
+      [HttpPost]
+      [ProducesResponseType(StatusCodes.Status201Created)]
+      [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+      public override async Task<IActionResult> Create([FromBody]CreateCareUserDTO dto)
+      {
+         if (dto is null)
+            throw new BadInputException("DTO is null", "Er is geen data meegegeven.");
+         _dateValidator.Validate(dto);
+
+         return await base.Create(dto);
       }
 
       [HttpPut("{id}")]
@@ -27,19 +42,17 @@ namespace Singer.Controllers
       public override async Task<IActionResult> Update(Guid id, [FromBody]UpdateCareUserDTO dto)
       {
          if (dto is null)
-         {
-            throw new BadInputException(nameof(dto));
-         }
+            throw new BadInputException("DTO is null", "Er is geen data meegegeven.");
+         var model = ModelState;
+         if (!model.IsValid)
+            throw new BadInputException("Invalid dto", $"De data is niet geldig: {model}");
+         _dateValidator.Validate(dto);
 
          if ((dto.LegalGuardianUsersToAdd?.Count ?? 0) > 0)
-         {
             await _careUserService.AddLinkedUsers(id, dto.LegalGuardianUsersToAdd);
-         }
 
          if ((dto.LegalGuardianUsersToRemove?.Count ?? 0) > 0)
-         {
             await _careUserService.RemoveLinkedUsers(id, dto.LegalGuardianUsersToRemove);
-         }
 
          var result = await DatabaseService.UpdateAsync(id, dto);
          return Ok(result);
